Pagepen | Finally a better solution to handle/highlight nested dom-nodes
This commit contains a proper first solution to highlighting text-content within nested dom nodes or partially selected ranges.
All in all, the key idea here is to use a Map which stores a parent node as the key and its value is an object with lists of both the original children nodes and also the newly created fragments which have our highlighted content.
This map gets filled by the tree-walker loop and once we're done looping, we call an updateDOM method and pass our updated or filled Map as an argument to this method.
The updateDOM method loops through the map and for each map key i.e. the parentNode, it iterates through the list of the original childnodes and replaces each child node, with the corresponding fragment that we created, which of course is also present in the same object alongside the original childnodes.
Keeping the DOM update separate from the initial loop also has a huge added benefit that the tree-walker can now easily traverse through the original dom without any hacks required from our end. All we do is keep a map, which has all the information needed to update the dom. Once the tree-walker is done with its loop, we can use the map to make the required updates.