Pagepen | Improvements and updates to the project structure.
This commit makes a few improvements to the project structure and adds some minor updates.
The updates are summarized below in brief:

The folders are organized as ui, contentScripts and backgroundScripts to better represent the structure of a browser extension.
The Webpack config has been updated as per the above point.
The extension manifest too has been updated to capture the changes.
A script has been added to generate TypeScript declaration files.