Pagepen | Update positioning of the Update toolbar
Place the Update toolbar based on the mouse click position instead of the end range position.
Add function getMouseClick position to place the toolbar as stated in the above point.
Update getPosition function to have a more descriptive name getRangeEndPosition
Update the css for UpdateToolbar so the toolbar is placed right above where the mouse click happens when clicking a highlighted text element.