feature/redesign: Add initial functional version of the redesign
This commit contains the basic redesign of the extension ui.
The redesign changes the way the extension is supposed to function even though the basic functionality and use case remains the same.
The redesign gets rid of the toolbars (initial design) as well as the comment box with multiple buttons like Add, Cancel, Colors etc. (the second version).
Instead, here the idea is display a simple minimal row of color options when some text is selected.
Currently a comment text area is also displayed below the color options. This may change in future and the comment box may only be shown as a part of the update flow instead of the create one.